Module Coq.Loader

val plugin_handler : (string list -> unit) option -> Mltop.PluginSpec.t -> unit

plugin_handler user_loader Plugin loader that will also load the instrumentation plugins for serlib if those available. user_loader can be used to override the default of Fl_dynload.load_packages as loader.