Module Message.Payload

Coq provides payload to our layer via two different mechanisms:

feedback messages

error exceptions

In both cases, the payload is the same, and it comes via different ways due to historical reasons. We abstract the payload as to better handle the common paths.

type 'l t = {

range : 'l option;

quickFix : 'l Lang.Qf.t list option;

msg : Pp.t;

}

val make : ?range:'l -> ?quickFix:'l Lang.Qf.t list -> Pp.t -> 'l t

val map : f:('l -> 'm) -> 'l t -> 'm t