Module Protect.Error

This module reifies Coq side effects into an algebraic structure.

This is very convenient for upper layer programming.

As of today we handle feedback, exceptions, and interruptions.

type 'l t = private

| User of 'l Message.Payload.t

| Anomaly of 'l Message.Payload.t