Module Fleche.Config

type t = {

mem_stats : bool;
(*
mem_stats Call Obj.reachable_words for every sentence. This is very slow and not very useful, so disabled by default
*)

gc_quick_stats : bool;
(*
gc_quick_stats Show the diff of Gc.quick_stats data for each sentence
*)

eager_diagnostics : bool;
(*
eager_diagnostics Send (full) diagnostics after processing each
*)

ok_diagnostics : bool;
(*
Show diagnostic for OK lines
*)

goal_after_tactic : bool;
(*
When showing goals and the cursor is in a tactic, if false, show goals before executing the tactic, if true, show goals after
*)

show_coq_info_messages : bool;
(*
Ignore `msg_info` messages in diagnostics
*)

}

val default : t

val v : t Stdlib.ref