Module Fleche.Contents

type t = private {

raw : string;
(*
That's the original, unprocessed document text, including markdown
*)

text : string;
(*
That's the text to be sent to the prover, already processed, and stripped from markdown, encoded in UTF-8
*)

last : Lang.Point.t;
(*
Last point of text, you can derive n_lines from here
*)

lines : string Stdlib.Array.t;
(*
text split in lines
*)

}

module R : sig ... end

val make : uri:Lang.LUri.File.t -> raw:string -> t R.t