Module Doc.Env

Enviroment external to the document, this includes for now the init Coq state and the workspace, which are used to build the first state of the document, usually by importing the prelude and other libs implicitly.

type t = private {

init : Coq.State.t;

workspace : Coq.Workspace.t;

files : Coq.Files.t;

}

val make : init:Coq.State.t -> workspace:Coq.Workspace.t -> files:Coq.Files.t -> t

val inject_requires : extra_requires:Coq.Workspace.Require.t list -> t -> t