Module Info.Offset

type t = int

val in_range : ?range:Lang.Range.t -> t -> bool

val gt_range : ?range:Lang.Range.t -> t -> bool

type offset_table = string

to_offset will help to resolve a position from for example (line,col) to an offset, but in some case requires a lookup method.

val to_offset : t -> offset_table -> int

val to_string : t -> string