Module Fleche.Stats

module Kind : sig ... end

time and memory-based stats

type t = {

time : float;

memory : float;

}

val record : kind:Kind.t -> f:('a -> 'b) -> 'a -> 'b * t

record ~kind ~f x returns f x with timing and memory use data attached to it; it will also update the global table for kind

val get_accumulated : kind:Kind.t -> t

get_accumulated ~kind returns global accumulated stats for kind

val reset : unit -> unit

reset () Reset global accumulated stats

module Global : sig ... end

val pp_words : Stdlib.Format.formatter -> float -> unit

Pretty-print memory info as words