Module Theory.Check

val maybe_check : io:Io.CallBack.t -> token:Coq.Limits.Token.t -> (Int.Set.t * Doc.t) option

Check pending documents, return None if there is none pending, or Some rqs the list of requests ready to execute after the check. Sends progress and diagnostics notifications using output function ofn.

val set_scheduler_hint : uri:Lang.LUri.File.t -> point:(int * int) -> unit