Module Theory.Register

module InjectRequire : sig ... end

List of additional includes to inject into a document

module Completed : sig ... end

Run an action when a document has completed checking, attention, with or without errors.