Module Theory.Request

type request =

| FullDoc of {

uri : Lang.LUri.File.t;

}

| PosInDoc of {

uri : Lang.LUri.File.t;

point : int * int;

version : int option;

postpone : bool;

}

type t = {

id : int;

request : request;

}

type action =

| Now of Doc.t

| Postpone

| Cancel

val add : t -> action

Add a request to be served; returns Postpone if request is added to the queue, Now doc if the request is available. Cancel means "we will never be able to serve this"

val remove : t -> unit

Removes the request from the list of things to wake up