Module Fleche.Types

module Point : sig ... end

ATTENTION: character is a Unicode caracter position, thus from Coq that usually requires conversion, as it will report the column offset in bytes. But offset is in bytes for now, as our downstream clients prefer this format.

module Range : sig ... end

module Diagnostic : sig ... end