Module Fleche_lsp.Io

JSON-RPC input/output

val read_message : Stdlib.in_channel -> (Base.Message.t, string) Result.t option

Read a JSON-RPC message from channel; None signals EOF

val send_message : Stdlib.Format.formatter -> Base.Message.t -> unit

Send a JSON-RPC message to channel

val set_log_fn : (string -> Yojson.Safe.t -> unit) -> unit

Set a debug function log action object which is called for read and send, useful to do low-level protocol debug