Module Diagnostic.Mode

type t =

| String

| Pp

Flèche diagnostics store the message as a Pp.t box format, but usually LSP standard mandates the message field to be a string, thus we allow clients to select the mode.

val set : t -> unit