Module Point.Mode

type t =

| LineColumn
(*
Points are standard LSP objects with line character field; this is the default
*)

| Offset
(*
Points are objects with only the offset
*)

| Full
(*
Points include / require line, character, and offset field
*)

Internally, Flèche always works with a point type which includes both line/column and offset information. However, this is beyond the LSP standard, so clients can configure the input / output behavior here

val set : t -> unit

Set the mode for serialization.