Module Lang.Point

type t = {

line : int;

character : int;

offset : int;

}

ATTENTION: character is a Unicode caracter position, thus from Coq that usually requires conversion, as it will report the column offset in bytes. But offset is in bytes for now, as our downstream clients prefer this format.

val pp : Stdlib.Format.formatter -> t -> unit