Module Petanque_shell.Protocol_shell

module Lsp = Fleche_lsp

module SetWorkspace : sig ... end

set_workspace { debug; root } sets the current workspace to the directory specified in root

module TableOfContents : sig ... end

toc { uri } returns the table of contents for a document; the semantics are quite Coq-specific, in particular, each sentence of the document can contribute