Module Serlib.Ser_environ

val abstract_env : bool Stdlib.ref

Control whether to output environs, they are huge, and in some case problematic; if abstract_env is true, they will be serialized as abstract

type env = Environ.env

val env_of_sexp : Sexplib.Sexp.t -> env

val sexp_of_env : env -> Sexplib.Sexp.t

type ('constr, 'types) punsafe_judgment = ('constr, 'types) Environ.punsafe_judgment

val punsafe_judgment_of_sexp : (Sexplib.Sexp.t -> 'constr) -> (Sexplib.Sexp.t -> 'types) -> Sexplib.Sexp.t -> ('constr, 'types) punsafe_judgment

val sexp_of_punsafe_judgment : ('constr -> Sexplib.Sexp.t) -> ('types -> Sexplib.Sexp.t) -> ('constr, 'types) punsafe_judgment -> Sexplib.Sexp.t

type unsafe_judgment = Environ.unsafe_judgment

val unsafe_judgment_of_sexp : Sexplib.Sexp.t -> unsafe_judgment

val sexp_of_unsafe_judgment : unsafe_judgment -> Sexplib.Sexp.t