Module Fleche.Config

module Unicode_completion : sig ... end

type t = {

mem_stats : bool;
(*
mem_stats Call Obj.reachable_words for every sentence. This is very slow and not very useful, so disabled by default
*)

gc_quick_stats : bool;
(*
gc_quick_stats Show the diff of Gc.quick_stats data for each sentence
*)

client_version : string;

eager_diagnostics : bool;
(*
eager_diagnostics Send (full) diagnostics after processing each
*)

goal_after_tactic : bool;
(*
When showing goals and the cursor is in a tactic, if false, show goals before executing the tactic, if true, show goals after
*)

show_coq_info_messages : bool;
(*
Show `msg_info` messages in diagnostics
*)

show_notices_as_diagnostics : bool;
(*
Show `msg_notice` messages in diagnostics
*)

admit_on_bad_qed : bool;
(*
admit_on_bad_qed There are two possible error recovery strategies when a Qed fails: one is not to add the constant to the state, the other one is admit it. We find the second behavior more useful, but YMMV.
*)

debug : bool;
(*
Enable debug on Coq side, including backtraces
*)

unicode_completion : Unicode_completion.t;

max_errors : int;

pp_type : int;
(*
Pretty-printing type in Info Panel Request, 0 = string; 1 = Pp.t; 2 = Coq Layout Engine
*)

show_stats_on_hover : bool;
(*
Show stats on hover
*)

show_loc_info_on_hover : bool;
(*
Show loc info on hover
*)

show_universes_on_hover : bool;
(*
Show universe data on hover
*)

show_state_hash_on_hover : bool;
(*
Show state hash on hover, useful for debugging
*)

pp_json : bool;
(*
Whether to pretty print the protocol JSON on the wire
*)

send_perf_data : bool;
(*
Whether to send the perf data notification
*)

send_diags : bool;
(*
Whether to send the diagnostics notification
*)

verbosity : int;
(*
Verbosity, 1 = reduced, 2 = default. As of today reduced will disable all logging, and the diagnostics and perf_data notification
*)

check_only_on_request : bool;
(*
Experimental setting to check document lazily
*)

send_diags_extra_data : bool;
(*
Send extra diagnostic data on the `data` diagnostic field.
*)

send_serverStatus : bool;
(*
Send server status client notification to the client
*)

}

val default : t

val v : t Stdlib.ref