Module Theory.Check

val maybe_check : io:Io.CallBack.t -> (Int.Set.t * Doc.t) option

Check pending documents, return None if there is none pending, or Some rqs the list of requests ready to execute after the check. Sends progress and diagnostics notifications using output function ofn.